Read to understand BookWood for comics
Need help to develop our BookWood for comic version. Let´s imagine a tablet, the 2/3 upper section of the screen is the
3D section. In this section, we will display Sketch up files. Each scene will have a flag to synchronize text and scenes.
A scene is defined by the author of the comic. It is a still image with the characters of the story that are present 
at that moment in the history.
The 1/3 bottom of the screen is the text window. Here will display the text with hidden flags so to synchronize text 
and scenes.
The synchronization must work in both ways, if the reader goes down the text or if he reader changes the scene view 
in the upper window. The reader must have full control just like when you are seeing 3D buildings in a city.
Thanks for your help!